Mobile resolution - Checkboxes are not right-aligned in the Application app's grid #5235
Moved common handlers to TreeGrid from its children.
